Edited experiment area and added gui file
I have finally manged to create a function which generates a board.
This function is still in the experiment file. It is improvable
and also the function check() in the function file needs to be made scalable
with n.
I have created the file gui to play around with tkinker.